refactor: moved submission-wizard steps to /steps folder
