+++ app/components/pages/submission-wizard/SubmissionPage.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import { Route, Switch } from 'react-router-dom'
 import WithCurrentSubmission from './WithCurrentSubmission'
-import AuthorDetailsPage from './Author/AuthorDetailsPage'
-import FileUploadsPage from './Files/FileUploadsPage'
-import ManuscriptMetadata from './Submission/ManuscriptMetadata'
-import ReviewerSuggestions from './Editors/ReviewerSuggestions'
-import { schema as fileUploadsSchema } from './Files/FileUploadsSchema'
-import { schema as authorDetailsSchema } from './Author/AuthorDetailsSchema'
-import { schema as manuscriptMetadataSchema } from './Submission/ManuscriptMetadataSchema'
-import { schema as reviewerSuggestionsSchema } from './Editors/ReviewerSuggestionsSchema'
+import AuthorDetailsPage from './steps/Author/AuthorDetailsPage'
+import FileUploadsPage from './steps/Files/FileUploadsPage'
+import ManuscriptMetadata from './steps/Submission/ManuscriptMetadata'
+import ReviewerSuggestions from './steps/Editors/ReviewerSuggestions'
+import { schema as fileUploadsSchema } from './steps/Files/FileUploadsSchema'
+import { schema as authorDetailsSchema } from './steps/Author/AuthorDetailsSchema'
+import { schema as manuscriptMetadataSchema } from './steps/Submission/ManuscriptMetadataSchema'
+import { schema as reviewerSuggestionsSchema } from './steps/Editors/ReviewerSuggestionsSchema'
 import WizardStep from './WizardStep'
 
 const SubmissionPage = ({ match, history }) => (
