+++ app/components/pages/submission-wizard/steps/Author/AuthorDetails.md
