+++ app/components/pages/submission-wizard/steps/Editors/ReviewerSuggestionsSchema.js
