+++ app/components/pages/submission-wizard/steps/Files/FileUploads.js
@@ -3,13 +3,13 @@
 import React from 'react'
 import { Box } from 'grid-styled'
 import ConfigurableEditor from 'xpub-edit/src/components/configurable/ConfigurableEditor'
-import FileUpload from '../../../ui/molecules/FileUpload'
-import ValidatedField from '../../../ui/atoms/ValidatedField'
-import IconTextBold from '../../../ui/atoms/icons/TextBold'
-import IconTextItalic from '../../../ui/atoms/icons/TextItalic'
-import IconTextUnderline from '../../../ui/atoms/icons/TextUnderline'
-import IconTextSub from '../../../ui/atoms/icons/TextSub'
-import IconTextSup from '../../../ui/atoms/icons/TextSup'
+import FileUpload from '../../../../ui/molecules/FileUpload'
+import ValidatedField from '../../../../ui/atoms/ValidatedField'
+import IconTextBold from '../../../../ui/atoms/icons/TextBold'
+import IconTextItalic from '../../../../ui/atoms/icons/TextItalic'
+import IconTextUnderline from '../../../../ui/atoms/icons/TextUnderline'
+import IconTextSub from '../../../../ui/atoms/icons/TextSub'
+import IconTextSup from '../../../../ui/atoms/icons/TextSup'
 
 // order of props affects order of menu buttons
 const Editor = ({ validationStatus, ...props }) => (
