+++ app/components/pages/submission-wizard/steps/Files/FileUploadsSchema.js
