+++ app/components/pages/submission-wizard/steps/Submission/ManuscriptMetadata.md
