+++ app/components/pages/submission-wizard/steps/Submission/ManuscriptMetadataSchema.js
